Implement Google OAuth2 Sign In
This commit implements a third party sign in pipeline that allows
users to sign in *only* if they have a User account created for them
with a matching email address.

In the case the admin user hasn't been created, the user is redirected
to a page alerting the user to the issue. It also provides a link
that can be used to resume the authentication (without needing to
log back in to the third party system).