CSV sanity pre-check, initial auth flow check fix
CSVs now get checked for duplicate columns. The problem arising
with duplicate column headers is this:

We don't control remote headers. If there are duplicates, we could
rename them on our side, but then we'd need to store the index of
the column and make sure the end user doesn't change it.

If we can ensure there are no duplicate columns, then users can change
them around. We've assumed they will want to do this because we
display the columns in the UI by the first five.