Consolidate importing, add updater manage.py cmd
Signals loading is fixed and importing/updating from
existing models is now consolidated into a DynamicModel
method. This is used post-refine to update and also
used in a management command:

    ./manage.py refresh_data_sources --name "Brandon's Sheet"

Again, note that this only updates existing models. Any
changes to the sheet's columns will be ignored. Users need
to go through the refine-and-import wizard to make these types
of changes.