Sorting and binding mostly work together
One remaining issue: Add new works if you haven't re-ordered
anything. But once you've re-ordered and then add new, the new
element gets pushed one up from the last item, as opposed to
actually at the end of the list. This is despite the fact that
the data model actually has the correct ordering, so this may
be a race condition between the two modules.