Meta updates to reconcile frontend and backend
This changes the handling of choice fields on the frontend (we now
pre-parse the choice field so they're the same on both sides).

Short-text fields are removed because nobody likes setting max_length.
In the list admin, we will only show an input field. The detail
view will still have textareas, though. But our assumption is that
few will use that.