Use Django 2.2.7 due to 2.2.8 breaking change
The 2.2.8 fix for the following issue causes collaborate
to not display the inline models as editable:

CVE-2019-19118: Privilege escalation in the Django admin.

Collaborate isn't vulnerable to this problem because we
don't have any signals that could possibly escalate privs
when saving any dynamic models. I'll need to find a way
to explicitly make this work with 2.2.8, but for now I'm
pushing a fix to get Collaborate working as expected.