Implement max records for avoiding request timeouts
Large imports will cause Collaborate deployments to timeout,
returning 504 errors to the end-user. This is confusing
because the records can ultimately import. To get around this,
we add a new settings variable in collaborate/settings.py
to control maximum in-request imports:

    MAX_IMPORT_RECORDS=750

You can set this to None or 0 to remove this functionality.
The background refresher task's behavior isn't changed by this,
so for large imports, we import a smaller chunk initially in
the request thread and then rely on the refresh_data_sources
management command to import full datasets:

  ./manage.py refresh_data_sources