Heroku clock: reduce memory usage
This reduces the total number of parallel queue executors
available to the Heroku scheduled clock (background importer).
While not a memory leak, the large pool of initialized executors
will eat up a significant amount of memory, quickly overloading
the low amount of RAM available to the base level Heroku instance.